requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.11.8
chardet==3.0.4
dnspython==1.16.0
idna==2.10
oauthlib==3.1.0
pymongo==3.11.2
python-dotenv==0.15.0
requests==2.25.0
requests-oauthlib==1.3.0
soupsieve==2.0.1
urllib3==1.26.2
